requirements.txt
openai==1.59.6
httpx==0.28.1
python-dotenv==1.0.1
langchain==0.3.14
langgraph==0.2.61
